split Elasticsearch integration back into separate v5 and v6 integrations
